+++ LookupAnything/Framework/Lookups/Buildings/BuildingLookupProvider.cs
@@ -25,12 +25,13 @@ internal class BuildingLookupProvider : BaseLookupProvider
     ** Public methods
     *********/
     /// <summary>Construct an instance.</summary>
+    /// <param name="monitor">Encapsulates monitoring and logging.</param>
     /// <param name="reflection">Simplifies access to private game code.</param>
     /// <param name="gameHelper">Provides utility methods for interacting with the game code.</param>
     /// <param name="config">The mod configuration.</param>
     /// <param name="codex">Provides subject entries.</param>
-    public BuildingLookupProvider(IReflectionHelper reflection, GameHelper gameHelper, Func<ModConfig> config, ISubjectRegistry codex)
-        : base(reflection, gameHelper)
+    public BuildingLookupProvider(IMonitor monitor, IReflectionHelper reflection, GameHelper gameHelper, Func<ModConfig> config, ISubjectRegistry codex)
+        : base(monitor, reflection, gameHelper)
     {
         this.Config = config;
         this.Codex = codex;