+++ LookupAnything/Framework/Lookups/Tiles/TileLookupProvider.cs
@@ -28,12 +28,13 @@ internal class TileLookupProvider : BaseLookupProvider
     ** Public methods
     *********/
     /// <summary>Construct an instance.</summary>
+    /// <param name="monitor">Encapsulates monitoring and logging.</param>
     /// <param name="reflection">Simplifies access to private game code.</param>
     /// <param name="gameHelper">Provides utility methods for interacting with the game code.</param>
     /// <param name="config">The mod configuration.</param>
     /// <param name="showRawTileInfo">Whether to show raw tile info like tilesheets and tile indexes.</param>
-    public TileLookupProvider(IReflectionHelper reflection, GameHelper gameHelper, Func<ModConfig> config, Func<bool> showRawTileInfo)
-        : base(reflection, gameHelper)
+    public TileLookupProvider(IMonitor monitor, IReflectionHelper reflection, GameHelper gameHelper, Func<ModConfig> config, Func<bool> showRawTileInfo)
+        : base(monitor, reflection, gameHelper)
     {
         this.Config = config;
         this.ShowRawTileInfo = showRawTileInfo;